% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarDiff.R
\name{covarDiff}
\alias{covarDiff}
\title{Compute the subject-level covariate difference for covariate adjustment}
\usage{
covarDiff(
  meta,
  subject_name,
  covariate_name,
  ctrl_subjs,
  case_subjs,
  id_ctrl,
  id_case,
  scale = 1
)
}
\arguments{
\item{meta}{a data frame of meta information}

\item{subject_name}{a character for subject name in meta}

\item{covariate_name}{a character for covariate name in meta}

\item{ctrl_subjs}{a vector of character for all control subjects}

\item{case_subjs}{a vector of character for all case subjects}

\item{id_ctrl}{index of control subject to compare}

\item{id_case}{index of case subject to compare}

\item{scale}{scale for covariate difference. If it is NULL, we will scale the covariate difference by its standard deviation.}
}
\value{
a vector of numeric values representing the covariate differences
}
\description{
Compute the subject-level covariate difference for covariate adjustment
}
