% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarDiffDummy.R
\name{covarDiffDummy}
\alias{covarDiffDummy}
\title{Compute the subject-level categorical covariate difference for covariate adjustment}
\usage{
covarDiffDummy(
  meta,
  subject_name,
  covariate_name,
  ref_name = NULL,
  ctrl_subjs,
  case_subjs,
  id_ctrl,
  id_case
)
}
\arguments{
\item{meta}{a data frame of meta information}

\item{subject_name}{a character for subject name in meta}

\item{covariate_name}{a character for covariate name in meta}

\item{ref_name}{a character representing the reference level for the categorical covariate}

\item{ctrl_subjs}{a vector of character for all control subjects}

\item{case_subjs}{a vector of character for all case subjects}

\item{id_ctrl}{index of control subject to compare}

\item{id_case}{index of case subject to compare}
}
\value{
a data frame of covariate difference between a category and the reference category
}
\description{
Compute the subject-level categorical covariate difference for covariate adjustment
}
