% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize the single-cell count matrix}
\usage{
normalize(count, meta, option = "none")
}
\arguments{
\item{count}{raw count matrix}

\item{meta}{meta information}

\item{option}{a character value representing the normalization method to apply to the count matrix}
}
\value{
normalized count matrix
}
\description{
Normalize the single-cell count matrix
}
