% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize the single-cell count matrix}
\usage{
normalize_data(count, meta, subject_name, option = "none")
}
\arguments{
\item{count}{raw count matrix.}

\item{meta}{meta information.}

\item{subject_name}{a character for subject name in \code{meta}.}

\item{option}{a character value representing the normalization method to apply to the count matrix.}
}
\value{
normalized count matrix
}
\description{
Normalize the single-cell count matrix
}
