% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dij_by_covar.R
\name{plot_dij_by_covar}
\alias{plot_dij_by_covar}
\title{Make a scatterplot to contrast the covariate difference and subject-level pairwise difference}
\usage{
plot_dij_by_covar(
  count,
  meta,
  res_table = NULL,
  normalize_option = "pooling",
  cond_name,
  subject_name,
  ctrl_cond,
  case_cond,
  feature,
  covariate_name,
  xlab = paste0(covariate_name, "_diff"),
  ylab = "d_ij",
  title = feature
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{a data frame of meta information.}

\item{res_table}{result table of ULV analysis.}

\item{normalize_option}{a character value to show the normalization method to
apply to the count matrix.}

\item{cond_name}{a character for condition name in \code{meta}.}

\item{subject_name}{a character for subject name in \code{meta}.}

\item{ctrl_cond}{a character for control name.}

\item{case_cond}{a character for case name.}

\item{feature}{a character for the feature to plot.}

\item{covariate_name}{a character for covariate name in \code{meta}.}

\item{xlab}{text label for x-axis.}

\item{ylab}{text label for y-axis.}

\item{title}{title of the plot.}
}
\value{

}
\description{
Make a scatterplot to contrast the covariate difference and subject-level pairwise difference
}
