% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_by_covar.R
\name{plot_dist_by_covar}
\alias{plot_dist_by_covar}
\title{Create boxplot of feature and subject-level summary statistic metric with covariate information}
\usage{
plot_dist_by_covar(
  count,
  meta,
  res_table = NULL,
  normalize_option = "pooling",
  cond_name,
  subj_name,
  feature,
  covar_name,
  sub_cond = NULL,
  show_subj = FALSE,
  xlab = covar_name,
  ylab = NULL,
  title = feature
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{a data frame of meta information.}

\item{res_table}{result table of ULV analysis.}

\item{normalize_option}{a character value to show the normalization method to
apply to the count matrix.}

\item{cond_name}{a character for condition name in \code{meta}.}

\item{subj_name}{a character for subject name in \code{meta}.}

\item{feature}{a character for the feature to plot.}

\item{covar_name}{a character for covariate name in \code{meta}.}

\item{sub_cond}{a vector of conditions to plot.}

\item{show_subj}{whether or not to show the subject ID on x-axis.}

\item{xlab}{text label on x-axis of the both plots.}

\item{ylab}{text label on y-axis of the boxplot for feature.}

\item{title}{title of the plot. Set to \code{feature} if not specified.}
}
\value{

}
\description{
Create boxplot of feature and subject-level summary statistic metric with covariate information
}
