% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_subj_dist.R
\name{plot_multi_subj_dist}
\alias{plot_multi_subj_dist}
\title{Create boxplot of feature from multiple subjects}
\usage{
plot_multi_subj_dist(
  count,
  meta,
  normalize_option = "none",
  cond_name,
  subj_name,
  feature,
  sub_cond = NULL,
  base_size = 15,
  y_range = NULL,
  show_subj = FALSE,
  xlab = "Subject",
  ylab = NULL,
  title = feature
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{a data frame of meta information.}

\item{normalize_option}{a character value to show the normalization method to
apply to the count matrix.}

\item{cond_name}{a character for condition name in \code{meta}.}

\item{subj_name}{a character for subject name in \code{meta}.}

\item{feature}{a character for the feature to plot.}

\item{sub_cond}{a vector of conditions to plot.}

\item{base_size}{base size of the figure. Set the default to 15.}

\item{y_range}{The range of y-axis in the plot. Set to NULL if not specified.}

\item{show_subj}{whether or not to show the subject ID on x-axis.}

\item{xlab}{text label on x-axis of the plot.}

\item{ylab}{text label on y-axis of the plot.}

\item{title}{title of the plot. Set to \code{feature} if not specified.}
}
\value{

}
\description{
\code{plot_multi_subj_dist} takes the count matrix and cell-level metadata,
 and creates a boxplot to show the distribution of \code{feature}, grouped by
 condition and subject.
}
\examples{
data('example_data')
count = example_data$count_matrix
count = count[1:10,]
meta = example_data$metadata
feat = rownames(count)[1]

plot_multi_subj_dist(count, meta, normalize_option = 'none',
                    cond_name = 'group_per_sample', subj_name = 'donor',
                    feature = feat)
}
