% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create volcano plots}
\usage{
plot_volcano(
  res_table,
  alpha = 0.1,
  PI_thres = 0.05,
  title = "Volcano Plot",
  xlab = "Estimated probabilistic index",
  ylab = "-log10(adjusted p-value)",
  add_labels = TRUE,
  max.overlaps = 15
)
}
\arguments{
\item{res_table}{a data frame with the column 'PI' representing the
probabilistic index, and the column 'padj' representing the adjusted
p-values of ULV analysis. The rownames should be the names of the features
(genes, proteins, etc.). The format is the same as the output from ULV
function.}

\item{alpha}{the threshold of adjusted p-values. The default value is set
as 0.1.}

\item{PI_thres}{the threshold of estimated probabilistic index.
The default value is set as 0.05.}

\item{title}{Title of the volcano plot.}

\item{xlab}{text label on x-axis of the plot.}

\item{ylab}{text label on y-axis of the plot.}

\item{add_labels}{whether or not to add text labels for the significant
features on the plot.}

\item{max.overlaps}{Equivalent of max.overlaps in ggrepel.
Set to 'Inf' to always display all labels.}
}
\value{
A ggplot2 figure.
}
\description{
\code{plot_volcano} create volcano plots. This function takes the ULV
analysis result table and produces volcano plots with the horizontal
axis representing the estimated probabilistic index.
}
\examples{
library(ULV)
data('example_data')
count = example_data$count_matrix
count = count[1:10,]
meta = example_data$metadata

res_table = fit_ULV(count, meta, normalize_option='pooling', subject_name = 'donor',
                    cond_name = 'group_per_sample', ctrl_cond = 'mild',
                    case_cond = 'severe', weighted = TRUE,
                    covariate_name_list=c('age_yr','sex'))
plot_volcano(res_table)

}
