% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_function.R
\name{run_DESeq}
\alias{run_DESeq}
\title{DESeq2}
\usage{
run_DESeq(
  count,
  meta,
  subject_name,
  cond_name,
  ctrl_cond,
  case_cond,
  covariate_name_list = NULL
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{meta data.}

\item{subject_name}{a character for the subject name in \code{meta}.}

\item{cond_name}{a character for the codition name in \code{meta}.}

\item{ctrl_cond}{a character for the control condition name.}

\item{case_cond}{a character for the case condition name.}

\item{covariate_name_list}{a vector of covariate name to adjust.
Set to NULL if there is no covariate adjustment.}
}
\value{
A result table for DESeq2 result.
}
\description{
DESeq2
}
