% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_function.R
\name{run_DESeq_sc}
\alias{run_DESeq_sc}
\title{DESeq2 at single cell level}
\usage{
run_DESeq_sc(
  count,
  meta,
  cond_name,
  ctrl_cond,
  case_cond,
  numerical_covar = NULL,
  categorical_covar = NULL
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{meta data.}

\item{cond_name}{a character for the codition name in \code{meta}.}

\item{ctrl_cond}{a character for the control condition name.}

\item{case_cond}{a character for the case condition name.}

\item{numerical_covar}{a vector of numerical covariate name to adjust.
Set to NULL if there is no covariate adjustment.}

\item{categorical_covar}{a vector of categorical covariate name to adjust.
Set to NULL if there is no covariate adjustment.}
}
\value{
A result table for DESeq2 result.
}
\description{
DESeq2 at single cell level
}
