% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_function.R
\name{run_MAST_glm}
\alias{run_MAST_glm}
\title{Fit MAST glm effect model}
\usage{
run_MAST_glm(
  count,
  meta,
  subj_name,
  cond_name,
  case_cond,
  ctrl_cond,
  categorical_covar = NULL,
  numerical_covar = NULL
)
}
\arguments{
\item{count}{Raw count matrix.}

\item{meta}{a data frame of metadata.}

\item{subj_name}{a character for the subject name in \code{meta}.}

\item{cond_name}{a character for the condition name in \code{meta}.}

\item{case_cond}{a character for case name.}

\item{ctrl_cond}{a character for control name.}

\item{categorical_covar}{a list of character values for the categorical covariate names}

\item{numerical_covar}{a list of character values for the numerical covariate names}
}
\value{
A data frame of MAST model fitting results containing log2-fold change, p-values, and BH-adjusted p-values.
}
\description{
Fit MAST glm effect model
}
