% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULV_function.R
\name{ULV}
\alias{ULV}
\title{ULV: a hybrid method for differential expression analysis using rank-base method and mixed-effect model}
\usage{
ULV(
  count,
  meta,
  normalize = TRUE,
  subject_name,
  cond_name,
  ctrl_cond,
  case_cond,
  weighted = FALSE,
  covariate_name_list = NULL
)
}
\arguments{
\item{count}{count matrix}

\item{meta}{a data frame of meta information}

\item{normalize}{if TRUE, then conduct the normalization during ULV.}

\item{subject_name}{a character for subject name in meta}

\item{cond_name}{a character for condition name in meta}

\item{ctrl_cond}{a character for control name}

\item{case_cond}{a character for case name}

\item{weighted}{a binary variable indicating whether the analysis is weighted by varying cluster sizes}

\item{covariate_name_list}{a vector of character of covariate names. If not, set it to NULL}
}
\value{
a result table summarizing the probabilistic index and p-value of each gene
}
\description{
ULV: a hybrid method for differential expression analysis using rank-base method and mixed-effect model
}
