% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_function.R
\name{wilcox_subject}
\alias{wilcox_subject}
\title{Wilcoxon test at subject level}
\usage{
wilcox_subject(
  count,
  meta,
  normalize_option = "pooling",
  subject_name,
  cond_name,
  case_cond,
  ctrl_cond,
  aggregate_method = 1
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{metadata.}

\item{normalize_option}{a character value to show the normalization method to
apply to the count matrix.}

\item{subject_name}{a character for subject name in \code{meta}.}

\item{cond_name}{a character for condition name in \code{meta}.}

\item{case_cond}{a character for case name.}

\item{ctrl_cond}{a character for control name.}

\item{aggregate_method}{aggregate method.}
}
\value{
A result table containing log2-fold change, p-values, and FDR-adjusted p-values.
}
\description{
Wilcoxon test at subject level
}
