% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize the single-cell count matrix}
\usage{
normalize(count, meta)
}
\arguments{
\item{count}{raw count matrix}

\item{meta}{meta information}
}
\value{
normalized count matrix
}
\description{
Normalize the single-cell count matrix
}
