% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_by_covar.R
\name{plot_dist_by_covar}
\alias{plot_dist_by_covar}
\title{Create boxplot of feature and subject-level summary statistic metric with covariate information}
\usage{
plot_dist_by_covar(
  count,
  meta,
  res_table = NULL,
  normalize_option = "pooling",
  cond_name,
  subj_name,
  feature,
  covar_name,
  sub_cond = NULL,
  xlab.p1 = covar_name,
  ylab.p1 = NULL,
  title.p1 = feature,
  xlab.p2 = covar_name,
  ylab.p2 = "Subject-level metric",
  title.p2 = "Subject-level metric"
)
}
\arguments{
\item{count}{count matrix.}

\item{meta}{a data frame of meta information.}

\item{res_table}{result table of ULV analysis.}

\item{normalize_option}{a character value to show the normalization method to
apply to the count matrix.}

\item{cond_name}{a character for condition name in \code{meta}.}

\item{subj_name}{a character for subject name in \code{meta}.}

\item{feature}{a character for the feature to plot.}

\item{covar_name}{a character for covariate name in \code{meta}.}

\item{sub_cond}{a vector of conditions to plot.}

\item{xlab.p1}{text label on x-axis of the first plot.}

\item{ylab.p1}{text label on y-axis of the first plot.}

\item{title.p1}{title of the first plot. Set to \code{feature} if not specified.}

\item{xlab.p2}{text label on x-axis of the second plot.}

\item{ylab.p2}{text label on y-axis of the second plot.}

\item{title.p2}{title of the second plot.}
}
\value{

}
\description{
Create boxplot of feature and subject-level summary statistic metric with covariate information
}
