% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULV_function.R
\name{ULV}
\alias{ULV}
\title{ULV: U-statistic based Latant Variable model that takes advantage of the robustness of rank-based methods and the statistical efficiency of parametric methods for small sample sizes}
\usage{
ULV(
  count,
  meta,
  normalize_option = "none",
  subject_name,
  cond_name,
  ctrl_cond,
  case_cond,
  weighted = FALSE,
  covariate_name_list = NULL
)
}
\arguments{
\item{count}{count matrix}

\item{meta}{a data frame of meta information}

\item{normalize_option}{a character value to show the normalization method to apply to the count matrix}

\item{subject_name}{a character for subject name in meta}

\item{cond_name}{a character for condition name in meta}

\item{ctrl_cond}{a character for control name}

\item{case_cond}{a character for case name}

\item{weighted}{a binary variable indicating whether the analysis is weighted by varying cluster sizes}

\item{covariate_name_list}{a vector of character of covariate names. If not, set it to NULL}
}
\value{
a result table summarizing the probabilistic index and p-value of each gene
}
\description{
ULV: U-statistic based Latant Variable model that takes advantage of the robustness of rank-based methods and the statistical efficiency of parametric methods for small sample sizes
}
\examples{
library(ULV)
data('example_data')
count = example_data$count_matrix
count = count[1:10,]
meta = example_data$metadata

res_table = ULV(count, meta, normalize_option='pooling', subject_name = 'donor',
                cond_name = 'group_per_sample', ctrl_cond = 'mild',
                case_cond = 'severe', weighted = TRUE,
                covariate_name_list=c('age_yr','sex'))
}
